test: move makeHistoryReviver test; move data to fixture module
